+++ test/config/test_preprocessor.py
@@ -15,6 +15,7 @@ import os
 from decimal import Decimal
 
 import pytest
+from librelane.config.variable import Instance, InstanceArray, Macro
 from pyfakefs.fake_filesystem_unittest import Patcher
 
 
@@ -88,6 +89,17 @@ def test_process_string():
     assert process_string("refg::$A", {"A": "B"}) == [
         "B"
     ], "refg:: without asterisks or ? did not return the same file path"
+
+
+def test_process_string_inline_expr():
+    from librelane.config.preprocessor import process_string
+
+    assert process_string("epic_sram_{X}_{Y}", {"X": 0, "Y": 0}) == "epic_sram_0_0"
+    assert process_string("{X}{Y}", {"X": 0, "Y": 0}) == "00"
+    assert (
+        process_string("-!!somecharacters{X}//--,,{Y}", {"X": "cool", "Y": "yeah"})
+        == "-!!somecharacterscool//--,,yeah"
+    )
 
 
 mmpt_raw = {
@@ -214,6 +226,125 @@ def test_preprocess_invalid_f_list():
     with pytest.raises(RuntimeError):
         preprocessed = preprocess_dict(
             config,
+            "/cwd",
+            pdk="sky130A",
+            pdkpath="/cwd",
+            scl="sky130_fd_sc_hd",
+        )
+
+def test_preprocess_array_macro_with_macro_instances():
+    from librelane.config.preprocessor import preprocess_dict
+
+    design = {
+        "meta": {"version": 2},
+        "PDK": "sky130A",
+        "STD_CELL_LIBRARY": "sky130_fd_sc_hd",
+        "DESIGN_NAME": "manual_macro_placement_test",
+        "VERILOG_FILES": "dir::src/*.v",
+        "MACROS": {
+            "spm": Macro(
+                gds="/foo",
+                lef="/foo",
+                instances={
+                    "epic_sram_512x8_{X}_{Y}": Instance(
+                        location=None,
+                        orientation="N",
+                        array=InstanceArray(
+                            offset=(100, 100), step=(100, 100), dimensions=(2, 2)
+                        ),
+                    )
+                },
+            )
+        },
+    }
+
+    preprocessed = preprocess_dict(
+        design,
+        "/cwd",
+        pdk="sky130A",
+        pdkpath="/cwd",
+        scl="sky130_fd_sc_hd",
+    )
+
+    expected = {
+        "meta": {"version": 2},
+        "PDK": "sky130A",
+        "STD_CELL_LIBRARY": "sky130_fd_sc_hd",
+        "DESIGN_NAME": "manual_macro_placement_test",
+        "VERILOG_FILES": ["/cwd/src/a_file.v", "/cwd/src/another_file.v"],
+        "DESIGN_DIR": "/cwd",
+        "PAD_CELL_LIBRARY": None,
+        "PDKPATH": "/cwd",
+        "MACROS": {
+            "spm": Macro(
+                gds="/foo",
+                lef="/foo",
+                instances={
+                    "epic_sram_512x8_0_0": Instance(
+                        location=[
+                            100.0,
+                            100.0,
+                        ],
+                        orientation="N",
+                    ),
+                    "epic_sram_512x8_1_0": Instance(
+                        location=[
+                            200.0,
+                            100.0,
+                        ],
+                        orientation="N",
+                    ),
+                    "epic_sram_512x8_0_1": Instance(
+                        location=[
+                            100.0,
+                            200.0,
+                        ],
+                        orientation="N",
+                    ),
+                    "epic_sram_512x8_1_1": Instance(
+                        location=[
+                            200.0,
+                            200.0,
+                        ],
+                        orientation="N",
+                    ),
+                },
+            )
+        },
+    }
+
+    assert preprocessed == expected, "Preprocessor produced a different result"
+
+
+def test_preprocess_array_macro_invalid():
+    from librelane.config.preprocessor import preprocess_dict
+
+    design = {
+        "meta": {"version": 2},
+        "PDK": "sky130A",
+        "STD_CELL_LIBRARY": "sky130_fd_sc_hd",
+        "DESIGN_NAME": "manual_macro_placement_test",
+        "VERILOG_FILES": "dir::src/*.v",
+        "MACROS": {
+            "spm": Macro(
+                gds="/foo",
+                lef="/foo",
+                instances={
+                    "epic_sram_512x8_{X}_{Y}": Instance(
+                        location=[1, 1],
+                        orientation="N",
+                        array=InstanceArray(
+                            offset=(100, 100), step=(100, 100), dimensions=(2, 2)
+                        ),
+                    )
+                },
+            )
+        },
+    }
+
+    with pytest.raises(RuntimeError):
+        preprocess_dict(
+            design,
             "/cwd",
             pdk="sky130A",
             pdkpath="/cwd",